+++ v2/src/transaction/blockchain/bsc/index.js
@@ -1,0 +1,1 @@
+// bsc
