+++ v2/src/transaction/blockchain/stellar/index.js
@@ -1,0 +1,1 @@
+// Stellar
